agentic_agents/requirements.txt
# SLIM Agent Communication
slim-bindings

# Google AI Integration
google-generativeai

# Environment Configuration
python-dotenv

# Standard libraries (built-in)
# asyncio
# json
# os
# sys
# time